backend/requirements.txt
# Core framework
fastapi==0.111.0
uvicorn==0.30.1

# OAuth & HTTP clients
httpx==0.27.0
requests==2.32.3
python-dotenv==1.0.1
starlette==0.26.0.post1
kombu==5.3.1

# Redis async support
redis==5.0.4

# Logging enhancements
logging==0.4.9.6

# Testing
pytest==8.2.1
pytest-asyncio==0.23.6
anyio==4.3.0

# Documentation generation (Sphinx)
sphinx==8.1.3
sphinx-autodoc-typehints==2.1.0

# Optional for development
python-multipart==0.0.9
